Refactor to reuse the same component to handle both cases: simple settings or regular settings
